requirements.txt
aiogram
python-dotenv
reportlab